fix(86): dictionaries as objects in response models
