+++ src/main/java/com/compstore/dto/dictionary/ProcessorBrandSimpleDTO.java
@@ -4,11 +4,9 @@ import java.util.UUID;
 import lombok.Data;
 
 @Data
-public class ProcessorBrandDTO {
+public class ProcessorBrandSimpleDTO {
 
     private UUID id;
 
     private String name;
-
-    private String processorBrandDeviceType;
 }
