+++ src/main/java/com/compstore/dto/dictionary/ScreenTypeDTO.java
@@ -4,11 +4,9 @@ import java.util.UUID;
 import lombok.Data;
 
 @Data
-public class ProcessorBrandDTO {
+public class ScreenTypeDTO {
 
     private UUID id;
 
     private String name;
-
-    private String processorBrandDeviceType;
 }
