+++ src/main/java/com/compstore/dto/pc/PCComboDataDTO.java
@@ -1,14 +1,14 @@
 package com.compstore.dto.pc;
 
 import com.compstore.dto.dictionary.GraphicsCardBrandDTO;
-import com.compstore.dto.dictionary.ProcessorBrandDTO;
+import com.compstore.dto.dictionary.ProcessorBrandSimpleDTO;
 import java.util.List;
 import lombok.Data;
 
 @Data
 public class PCComboDataDTO {
 
-    private List<ProcessorBrandDTO> processorBrands;
+    private List<ProcessorBrandSimpleDTO> processorBrands;
 
     private List<GraphicsCardBrandDTO> graphicsCardBrands;
