--- src/main/java/com/compstore/entity/ScreenTypeEntity.java
@@ -1,23 +1,0 @@
-package com.compstore.entity;
-
-import jakarta.persistence.*;
-import java.util.UUID;
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@Entity
-@Table(name = "screen_types")
-@NoArgsConstructor
-@AllArgsConstructor
-public class ScreenTypeEntity {
-
-    @Id @GeneratedValue private UUID id;
-
-    private String name;
-
-    @Column(name = "screen_type_device_type")
-    @Enumerated(value = EnumType.STRING)
-    private ScreenTypeDeviceType screenTypeDeviceType;
-}
