+++ src/main/java/com/compstore/entity/dictionary/enums/BrandDeviceType.java
@@ -1,4 +1,4 @@
-package com.compstore.entity;
+package com.compstore.entity.dictionary.enums;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
