+++ src/main/java/com/compstore/mapper/ProcessorBrandMapper.java
@@ -3,8 +3,8 @@ package com.compstore.mapper;
 import com.compstore.dto.dictionary.ProcessorBrandComboDataDTO;
 import com.compstore.dto.dictionary.ProcessorBrandCreateRequestDTO;
 import com.compstore.dto.dictionary.ProcessorBrandDTO;
-import com.compstore.entity.ProcessorBrandDeviceType;
-import com.compstore.entity.ProcessorBrandEntity;
+import com.compstore.entity.dictionary.ProcessorBrandEntity;
+import com.compstore.entity.dictionary.enums.ProcessorBrandDeviceType;
 import org.mapstruct.Mapper;
 import org.mapstruct.Mapping;
 import org.mapstruct.MappingTarget;
@@ -14,11 +14,15 @@ import org.mapstruct.ReportingPolicy;
 public interface ProcessorBrandMapper {
 
     @Mapping(target = "id", ignore = true)
+    @Mapping(target = "pcs", ignore = true)
+    @Mapping(target = "smartphones", ignore = true)
     ProcessorBrandEntity toEntity(ProcessorBrandCreateRequestDTO processorBrandCreateRequestDTO);
 
     ProcessorBrandDTO toDTO(ProcessorBrandEntity processorBrandEntity);
 
     @Mapping(target = "id", ignore = true)
+    @Mapping(target = "pcs", ignore = true)
+    @Mapping(target = "smartphones", ignore = true)
     void toEntity(
             ProcessorBrandCreateRequestDTO processorBrandUpdateRequest,
             @MappingTarget ProcessorBrandEntity existingProcessorBrand);
