+++ src/main/java/com/compstore/service/impl/PCServiceImpl.java
@@ -5,10 +5,10 @@ import com.compstore.dto.pc.PCCreateRequestDTO;
 import com.compstore.dto.pc.PCDTO;
 import com.compstore.dto.pc.PCFilteringRequestDTO;
 import com.compstore.dto.pc.PCFilteringResponseDTO;
-import com.compstore.entity.GraphicsCardBrandDeviceType;
-import com.compstore.entity.GraphicsCardBrandEntity;
-import com.compstore.entity.ProcessorBrandDeviceType;
-import com.compstore.entity.ProcessorBrandEntity;
+import com.compstore.entity.dictionary.GraphicsCardBrandEntity;
+import com.compstore.entity.dictionary.ProcessorBrandEntity;
+import com.compstore.entity.dictionary.enums.GraphicsCardBrandDeviceType;
+import com.compstore.entity.dictionary.enums.ProcessorBrandDeviceType;
 import com.compstore.entity.pc.*;
 import com.compstore.entity.pc.enums.DriveCapacity;
 import com.compstore.entity.pc.enums.DriveType;
