+++ src/main/java/com/compstore/service/impl/ProcessorBrandServiceImpl.java
@@ -3,8 +3,8 @@ package com.compstore.service.impl;
 import com.compstore.dto.dictionary.ProcessorBrandComboDataDTO;
 import com.compstore.dto.dictionary.ProcessorBrandCreateRequestDTO;
 import com.compstore.dto.dictionary.ProcessorBrandDTO;
-import com.compstore.entity.ProcessorBrandDeviceType;
-import com.compstore.entity.ProcessorBrandEntity;
+import com.compstore.entity.dictionary.ProcessorBrandEntity;
+import com.compstore.entity.dictionary.enums.ProcessorBrandDeviceType;
 import com.compstore.exception.NotFoundException;
 import com.compstore.mapper.ProcessorBrandMapper;
 import com.compstore.repository.ProcessorBrandRepository;
